refactor: 이미지 업로더 로직 커스텀훅으로 분리 후 적용 (#37)
